Fix the separation between reasoning and answer for Ollama
The separation was done wrong, saving the part that is BEFORE the <think> tag instead of the part AFTER the </think> closing tag. As a result, the message content was empty.

Note: this relies on the assumption that model reasoning always follows the pattern "<think> some reasoning </think> answer". To my knowledge this is the case everywhere.

Note 2: maybe it would be useful to have some kind of escape mechanism, as there are cases where this would break.