fix: bad separation between reasoning and answer for Ollama (#376)
The separation was done wrong, saving the part that is BEFORE the
<think> tag instead of the part AFTER the </think> closing tag. As a
result, the message content was empty.

Note: this relies on the assumption that model reasoning always follows
the pattern "\<think\> some reasoning \<\/think\> answer". To my
knowledge this is the case everywhere.

Note 2: maybe it would be useful to have some kind of escape mechanism,
as there are cases where this would break.

## Description
<!-- What does this PR do? -->


## PR Type

🐛 Bug Fix

## Checklist
- [x] I have added unit tests that prove my fix/feature works
- [x] New and existing tests pass locally
- [x] Documentation was updated where necessary
- [x] I have read and followed the [contribution
guidelines](https://github.com/mozilla-ai/any-llm/blob/main/CONTRIBUTING.md)```

I wrote a fix quickly because the issue in the logic is evident and it
didn't work in my real-use case. I haven't run the tests locally, and I
don't understand how this passed the reasoning integration test in the
first place.

---------

Co-authored-by: Nathan Brake <33383515+njbrake@users.noreply.github.com>
Co-authored-by: Nathan Brake <nathan@mozilla.ai>