fix(gemin): Implement `_verify_and_set_api_key`. (#410)
In #401 we stopped overriding `Provider.__init__` and thus we started to
rely again on the default `Provider._verify_and_set_api_key`.

We need to provide a custom implementation that checks for both API key
options instead of the plain string with a `/`.

- Added `test_gemini_initialization_with_env_var_api_key` and
`test_vertexai_initialization_with_env_var_api_key`.