feat: experimental AWS Sagemaker support (#411)
## Description
<!-- What does this PR do? -->
This PR adds experimental (untested) support for AWS Sagemaker. We don't
currently have credentials/access to test this
against a Sagemaker endpoint. Using the documentation and the existing
(and tested) implementation with AWS Bedrock, this is enough to create
what I think is a working implementation.

But until we get a first user/tester, I won't know for sure. I added a
logger.warning message to tell a user using sagemaker that our support
at this point is experimental, hopefully that's sufficient for the
moment.

## PR Type

<!-- Delete the types that don't apply --!>

🆕 New Feature

## Relevant issues

<!-- e.g. "Fixes #123" -->

## Checklist
- [x] I have added unit tests that prove my fix/feature works
- [x] New and existing tests pass locally
- [x] Documentation was updated where necessary
- [x] I have read and followed the [contribution
guidelines](https://github.com/mozilla-ai/any-llm/blob/main/CONTRIBUTING.md)```