feat: separate test for local providers (#422)
## Description
<!-- What does this PR do? -->
Moves testing of local providers to a separate integration test. It also
helped us catch a bug in the design (where we actually needed to have
OPENAI_API_KEY set, which should not be the case)

I only test llamafile and ollama. Testing for lmstudio and llamacpp
looks like it will take a bunch more effort that imo it's not worth at
this point, so I called it good with llamafile and ollama.

## PR Type

<!-- Delete the types that don't apply --!>

🆕 New Feature

💅 Refactor

🚦 Infrastructure

## Relevant issues

<!-- e.g. "Fixes #123" -->

## Checklist
- [x] I have added unit tests that prove my fix/feature works
- [x] New and existing tests pass locally
- [x] Documentation was updated where necessary
- [x] I have read and followed the [contribution
guidelines](https://github.com/mozilla-ai/any-llm/blob/main/CONTRIBUTING.md)```